+++ pip-wrapper/onyxdb/__init__.py
@@ -1,1 +1,4 @@
-__version__ = '0.1.0'
+from .client import OnyxDB
+
+__version__ = '0.2.0'
+__all__ = ['OnyxDB']